Make `Piece` immutable.
`Piece` is not a frozen dataclass which simplifies some of the code
and makes debugging/following the code logic simpler. In particular
it means the location information of a piece now has one source.

Tests and `testutils` have been updated accordingly with a new
format for `ssv` files introduced.

A new `Square` class has been added to facilitate cases where a
`Piece` with a coordinate is still the best solution.